refactored a bit to start
